refactor: pages and components folder structure
