+++ src/components/PokemonCard/lazyLoading/index.js
