+++ src/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import "./css/tailwind.css";
+import "assets/css/tailwind.css";
 import App from "./App";
 
 ReactDOM.render(
