+++ src/pages/PCP/components/Pokedex/PokedexList/index.css
