+++ src/pages/PCP/components/Pokedex/PokedexThumbnail/index.js.js
