+++ src/pages/PDP/components/PokemonDetails/About/Meterbar/index.js
