+++ src/pages/PDP/components/PokemonDetails/About/index.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState, useContext } from "react";
 import { PokemonDetailsContext } from "pages/PDP";
 import { hashedPokemonID } from "components/utils/helper/helper";
-import PokemonStats from "./components/PokemonStats";
+import PokemonStats from "./PokemonStats";
 
 function PokemonAbout() {
   const [pokemon, setPokemon] = useState([]);
